pyproject.toml
[project]
name = "linear-regression-waic"
description = "Bayesian Linear Regression and WAIC"
authors = [{ name = "杜 世橋 Du Shiqiao", email = "lucidfrontier.45@gmail.com" }]
license = { text = "MIT" }
dynamic = ["version"]
requires-python = ">=3.11,<3.13"
dependencies = [
    "numpyro>=0.13.2",
    "jaxtyping>=0.2.24",
    "arviz>=0.16.1",
    "cmdstanpy>=1.2.0",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.version]
source = "file"
path = "src/app/version.py"


[tool.pdm.scripts]
black = "black ."
pyright = "pyright ."
ruff_lint = "ruff ."
ruff_fix = "ruff --fix-only ."
test = "pytest tests --cov=app --cov-report=term --cov-report=xml"
format = { composite = ["ruff_fix", "black"] }
lint = { composite = ["ruff_lint", "pyright"] }
check = { composite = ["format", "lint", "test"] }

[tool.pdm.dev-dependencies]
dev = ["black>=23.11.0", "ruff>=0.1.7", "pyright>=1.1.339", "ipykernel>=6.27.1"]
[tool.pytest.ini_options]
filterwarnings = ["ignore::FutureWarning"]

[tool.black]
target-version = ["py311"]
# add directory names to exclude from analysis
extend-exclude = "deps"

[tool.ruff]
select = ["E", "F", "W"]
target-version = "py311"
ignore = ["F722"]
# add directory names to exclude from analysis
exclude = ["deps"]

[tool.ruff.per-file-ignores]
"*/__init__.py" = ['F401']

[tool.pyright]
pythonVersion = "3.11"
# add directory names to exclude from analysis
ignore = ["deps"]
